comment error

+++ src/main/java/com/sspu/community/controller/PublishController.java
@@ -28,7 +28,7 @@ public class PublishController {
     private QuestionService questionService;
 
     @GetMapping("/publish/{id}")
-    public String edit(@PathVariable(name = "id") Integer id,
+    public String edit(@PathVariable(name = "id") Long id,
                        Model model) {
         QuestionDTO question = questionService.getById(id);
         model.addAttribute("title", question.getTitle());
@@ -37,7 +37,6 @@ public class PublishController {
         model.addAttribute("id", question.getId());
         return "publish";
     }
-
 
 
     @GetMapping("/publish")
@@ -50,7 +49,7 @@ public class PublishController {
             @RequestParam(value = "title", required = false) String title,
             @RequestParam(value = "description", required = false) String description,
             @RequestParam(value = "tag", required = false) String tag,
-            @RequestParam(value = "id", required = false) Integer id,
+            @RequestParam(value = "id", required = false) Long id,
             HttpServletRequest request,
             Model model) {
         model.addAttribute("title", title);
@@ -69,7 +68,6 @@ public class PublishController {
             model.addAttribute("error", "标签不能为空");
             return "publish";
         }
-
         User user = (User) request.getSession().getAttribute("user");
         if (user == null) {
             model.addAttribute("error", "用户未登录");
